+++ log/query.go
@@ -265,11 +265,37 @@ func internalQuery(
 			if d.Details()&trace.QueryEvents == 0 {
 				return nil
 			}
-			ctx := with(*info.Context, TRACE, "ydb", "query", "do", "tx")
+			ctx := with(*info.Context, TRACE, "ydb", "query", "read", "row")
 			l.Log(ctx, "start")
 			start := time.Now()
 
 			return func(info trace.QueryReadRowDoneInfo) {
+				if info.Error == nil {
+					l.Log(ctx, "done",
+						latencyField(start),
+					)
+				} else {
+					lvl := ERROR
+					if !xerrors.IsYdb(info.Error) {
+						lvl = DEBUG
+					}
+					l.Log(WithLevel(ctx, lvl), "failed",
+						latencyField(start),
+						Error(info.Error),
+						versionField(),
+					)
+				}
+			}
+		},
+		OnReadResultSet: func(info trace.QueryReadResultSetStartInfo) func(trace.QueryReadResultSetDoneInfo) {
+			if d.Details()&trace.QueryEvents == 0 {
+				return nil
+			}
+			ctx := with(*info.Context, TRACE, "ydb", "query", "read", "rows")
+			l.Log(ctx, "start")
+			start := time.Now()
+
+			return func(info trace.QueryReadResultSetDoneInfo) {
 				if info.Error == nil {
 					l.Log(ctx, "done",
 						latencyField(start),